dma-buf: Fix the type of the return variable in dma_fence_wait_timeout()
After jiffies was widened, MAX_SCHEDULE_TIMEOUT gets truncated to
0xffffffff, which is sign-extended to a negative value.

(cherry picked from commit 173e06913498ee21c0cdac07e591dce89c68349d)

+++ drivers/dma-buf/dma-fence.c
@@ -168,7 +168,7 @@ dma_fence_signal(struct dma_fence *fence)
 signed long
 dma_fence_wait_timeout(struct dma_fence *fence, bool intr, signed long timeout)
 {
-	int rv;
+	signed long rv;
 
 	if (fence == NULL)
 		return (-EINVAL);
